+++ nomad/plan_normalization_test.go
@@ -62,5 +62,5 @@ func TestPlanNormalize(t *testing.T) {
 	}
 
 	optimizedLogSize := buf.Len()
-	assert.Less(t, float64(optimizedLogSize)/float64(unoptimizedLogSize), 0.65)
+	assert.Less(t, float64(optimizedLogSize)/float64(unoptimizedLogSize), 0.66)
 }